+++ src/components/dashboard/Spaces/types/spaces.d.ts
@@ -1,0 +1,15 @@
+export type Event = {
+  name: string
+  date: string //repalce with Date
+}
+
+export type Topic = {
+  name: string
+  posts: number
+}
+
+export type Stat = {
+  name: string
+  amount: number
+  icon: React.ReactNode
+}
